+++ src/main/java/chess/dto/request/CreateRoomDto.java
@@ -1,4 +1,4 @@
-package chess.dto;
+package chess.dto.request;
 
 public class CreateRoomDto {
