+++ src/main/java/chess/dto/response/RoomDto.java
@@ -1,21 +1,18 @@
-package chess.dto;
+package chess.dto.response;
 
 import chess.domain.Team;
-import chess.entity.Room;
 
 public class RoomDto {
 
     private final Long id;
     private final Team team;
     private final String title;
-    private final String password;
     private boolean status;
 
-    public RoomDto(Long id, Team team, String title, String password, boolean status) {
+    public RoomDto(Long id, Team team, String title, boolean status) {
         this.id = id;
         this.team = team;
         this.title = title;
-        this.password = password;
         this.status = status;
     }
 
@@ -31,15 +28,8 @@ public class RoomDto {
         return title;
     }
 
-    public String getPassword() {
-        return password;
-    }
-
     public boolean isStatus() {
         return status;
     }
 
-    public Room toRoom() {
-        return new Room(id, team, title, password, status);
-    }
 }
