+++ src/main/java/chess/dto/response/ScoreDto.java
@@ -1,4 +1,4 @@
-package chess.dto;
+package chess.dto.response;
 
 public class ScoreDto {
