+++ src/test/java/chess/service/ChessServiceTest.java
@@ -7,13 +7,13 @@ import chess.dao.BoardDao;
 import chess.dao.RoomDao;
 import chess.domain.Team;
 import chess.domain.state.BoardInitialize;
-import chess.dto.BoardDto;
-import chess.dto.CreateRoomDto;
-import chess.dto.GameStateDto;
-import chess.dto.PieceDto;
-import chess.dto.RoomDto;
-import chess.dto.ScoreDto;
-import chess.dto.StatusDto;
+import chess.dto.response.BoardDto;
+import chess.dto.request.CreateRoomDto;
+import chess.dto.response.GameStateDto;
+import chess.dto.response.PieceDto;
+import chess.dto.response.RoomDto;
+import chess.dto.response.ScoreDto;
+import chess.dto.response.StatusDto;
 import chess.entity.Room;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
